Restore previous clip_inplace hacks

+++ libEM/processor.cpp
@@ -6154,9 +6154,9 @@ void ToMassCenterProcessor::process_inplace(EMData * image)
 	int nz = image->get_zsize();
 
 	if (int_shift_only) {
-		float dx = -(floor(com[0] + 0.5f) - nx / 2);	// these used to be integers, but clip_inplace seems to be broken for dx<0, quick hack to avoid this routine
-		float dy = -(floor(com[1] + 0.5f) - ny / 2);
-		float dz = 0;
+		int dx = -(floor(com[0] + 0.5f) - nx / 2);
+		int dy = -(floor(com[1] + 0.5f) - ny / 2);
+		int dz = 0;
 		if (nz > 1) {
 			dz = -(floor(com[2] + 0.5f) - nz / 2);
 		}
